VoiceController.ktレスポンスクラスを別ファイルに切り出し #48
