+++ src/main/kotlin/com/assari/voicebooklm/presentation/controller/voice/VoiceController.kt
@@ -2,7 +2,6 @@ package com.assari.voicebooklm.presentation.controller.voice
 
 import com.assari.voicebooklm.presentation.controller.auth.ErrorResponse
 import com.assari.voicebooklm.usecase.memo.CreateMemoInput
-import com.assari.voicebooklm.usecase.memo.CreateMemoOutput
 import com.assari.voicebooklm.usecase.memo.CreateMemoUseCase
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
@@ -24,8 +23,6 @@ import org.springframework.web.server.ResponseStatusException
 import java.nio.file.Files
 import java.nio.file.Path
 import java.util.UUID
-import kotlin.time.Duration
-import kotlin.time.toJavaDuration
 
 /**
  * 音声アップロードを受け取り、メモ作成ユースケースを呼び出す
@@ -50,7 +47,7 @@ class VoiceController(
             ApiResponse(
                 responseCode = "201",
                 description = "メモ生成成功",
-                content = [Content(schema = Schema(implementation = CreateMemoResponse::class))],
+                content = [Content(schema = Schema(implementation = VoiceMemoCreatedResponse::class))],
             ),
             ApiResponse(
                 responseCode = "400",
@@ -73,7 +70,7 @@ class VoiceController(
         authentication: Authentication?,
         @RequestPart("file") file: MultipartFile,
         @RequestParam("language", required = false) language: String?,
-    ): ResponseEntity<CreateMemoResponse> {
+    ): ResponseEntity<VoiceMemoCreatedResponse> {
         val userId = extractUserId(authentication)
         validateFile(file)
         val tempPath = writeTempFile(file)
@@ -103,7 +100,7 @@ class VoiceController(
             voiceMemo.transcription.fallbackUsed,
             voiceMemo.formatting.fallbackUsed,
         )
-        val response = CreateMemoResponse.from(result)
+        val response = VoiceMemoCreatedResponse.from(result)
         return ResponseEntity.status(HttpStatus.CREATED).body(response)
     }
 
@@ -158,57 +155,3 @@ class VoiceController(
         }
     }
 }
-
-/**
- * メモ作成レスポンス
- */
-data class CreateMemoResponse(
-    val memoId: UUID,
-    val title: String,
-    val content: String,
-    val tags: List<String>,
-    val transcription: String?,
-    val transcriptionStatus: String,
-    val formattingStatus: String,
-    val processingTimeMillis: ProcessingTimeResponse,
-    val fallback: FallbackUsageResponse,
-) {
-    companion object {
-        fun from(result: CreateMemoOutput): CreateMemoResponse {
-            val voiceMemo = result.voiceMemo
-            return CreateMemoResponse(
-                memoId = voiceMemo.id,
-                title = voiceMemo.title ?: "",
-                content = voiceMemo.content ?: "",
-                tags = voiceMemo.tags,
-                transcription = voiceMemo.transcriptionText,
-                transcriptionStatus = voiceMemo.transcription.status.name,
-                formattingStatus = voiceMemo.formatting.status.name,
-                processingTimeMillis = ProcessingTimeResponse(
-                    transcription = result.processingTime.transcription.toMillis(),
-                    formatting = result.processingTime.formatting.toMillis(),
-                    persistence = result.processingTime.persistence.toMillis(),
-                    total = result.processingTime.total.toMillis(),
-                ),
-                fallback = FallbackUsageResponse(
-                    transcription = voiceMemo.transcription.fallbackUsed,
-                    formatting = voiceMemo.formatting.fallbackUsed,
-                ),
-            )
-        }
-    }
-}
-
-data class ProcessingTimeResponse(
-    val transcription: Long,
-    val formatting: Long,
-    val persistence: Long,
-    val total: Long,
-)
-
-data class FallbackUsageResponse(
-    val transcription: Boolean,
-    val formatting: Boolean,
-)
-
-private fun Duration.toMillis(): Long = this.toJavaDuration().toMillis()
